requirements.txt
# Kronos Trading System Requirements
# ===================================

# Core ML/Data
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Deep Learning (for Kronos)
torch>=2.0.0
einops>=0.8.0
safetensors>=0.4.0

# Gradient Boosting
xgboost>=2.0.0
lightgbm>=4.0.0

# Data Collection
aiohttp>=3.9.0
requests>=2.31.0
websockets>=12.0

# Exchange APIs
eth-account>=0.11.0  # For Hyperliquid wallet signing

# Data Storage
pyarrow>=14.0.0
fastparquet>=2023.10.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0

# Progress/Logging
tqdm>=4.66.0
colorlog>=6.8.0

# Utilities
huggingface_hub>=0.20.0
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Type checking
mypy>=1.7.0

# Development
black>=23.12.0
isort>=5.13.0
flake8>=7.0.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.27.0
